+++ CHANGELOG.md
@@ -29,3 +29,12 @@
 - cleaned up render button
 - cleaned up JPInterface and Canvas
 
+## [0.0.3]
+### Added
+- Generator class (SwingWorker) for calculating the fractal asynchronously  
+- Added MouseListener to Canvas and fixed mirror bug (retained java axis orientation, might change this later)
+
+### Changed
+- Moved all swing components to EDT
+- Moved fractal calculation to SwingWorker
+- Fixed JProgressBar and progressMessage in feedback panel